change: support using system packages with custom variant
Took a deeper look at #101

I think this change may be reasonable based on our discussion there.

Essentially, if `package` is not set in an entry in `_preload_library.json`, we will allow using the system-level package.

I am using this patch when creating my own variant, but figure it may be valuable/safe to upstream.

To address the note:

> I might add changes to allow users to use system CUDA at their own risk.
> (Though I don’t intend to change the default behavior)

I think this does this because all default builds are guaranteed to have `package` set in this file, so this is a noop for all the default/pypi-uploaded builds